added helper methods to all of these files. Added new file for method…

+++ src/test/java/com/sparta/owframework/OWWeatherDTO/Rain.java
@@ -5,16 +5,16 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class Rain{
 
 	@JsonProperty("3h")
-	private Integer jsonMember3h;
+	private Double jsonMember3h;
 
 	@JsonProperty("1h")
-	private Integer jsonMember1h;
+	private Double jsonMember1h;
 
-	public Integer getJsonMember3h(){
+	public Double getJsonMember3h(){
 		return jsonMember3h;
 	}
 
-	public Integer getJsonMember1h(){
+	public Double getJsonMember1h(){
 		return jsonMember1h;
 	}
 }
